refactor, move Factory as static nested class
